TT-3381b Transcription in dialog
- also show transcription in dialog based on choice.
